requirements.txt
africastalking==2.0.2
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
datetime-truncate==1.1.1
diff-match-patch==20241021
Django==6.0.2
django-admin-charts==1.5.0
django-admin-interface==0.32.0
django-bower==5.2.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-import-export==4.4.0
django-memoize==2.3.1
django-multiselectfield==1.0.1
django-nvd3==0.10.1
djangorestframework==3.16.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
schema==0.7.8
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
